expr.RecordExpr: Defuse spread expressions
The commit changes the behavior of record spread expressions so that
incoming values are defused. In order to accomplish this defuse/downcast
had to be moved into vam/expr to avoid an import cycle.
