+++ runtime/vam/expr/function/function.go
@@ -51,7 +51,7 @@ func New(sctx *super.Context, name string, narg int) (expr.Function, error) {
 		argmax = 2
 		f = &DatePart{sctx}
 	case "defuse":
-		f = NewDefuse(sctx)
+		f = newDefuse(sctx)
 	case "downcast":
 		argmin = 2
 		argmax = 2
