+++ runtime/vam/expr/recordexpr.go
@@ -33,6 +33,7 @@ func NewRecordExpr(sctx *super.Context, elems []RecordElem) Evaluator {
 		sctx:         sctx,
 		elems:        elems,
 		fieldIndexes: map[string]int{},
+		defuse:       NewDefuse(sctx),
 	}
 }
 
@@ -44,6 +45,8 @@ type recordExpr struct {
 	fields       []super.Field
 	fieldIndexes map[string]int
 	fieldVecs    []vector.Any
+	// defuse is for defusing spread values.
+	defuse *Defuse
 }
 
 func (r *recordExpr) Eval(this vector.Any) vector.Any {
@@ -64,7 +67,7 @@ func (r *recordExpr) Eval(this vector.Any) vector.Any {
 				vec = vector.NewOptionSome(r.sctx, vec)
 			}
 		case *SpreadElem:
-			vec = elem.Expr.Eval(this)
+			vec = r.defuse.Eval(elem.Expr.Eval(this))
 		default:
 			panic(elem)
 		}
